moving all query parsing logic into its own helper. Tests to come
